Added JSX tests and refactored.
