+++ extensions/jsx.yaml
@@ -18,7 +18,7 @@ contexts: !merge
         - - meta_scope: meta.interpolation.js
           - meta_content_scope: source.js.embedded.jsx
           - match: '}'
-            scope: punctuation.definition.interpolation.begin.js
+            scope: punctuation.definition.interpolation.end.js
             pop: true
         - expression
 
@@ -30,12 +30,12 @@ contexts: !merge
     - include: else-pop
 
   jsx-meta:
-    - meta_scope: meta.jsx
+    - meta_scope: meta.jsx.js
     - include: immediately-pop
 
   jsx-tag:
     - match: '<'
-      scope: meta.tag.js punctuation.definition.tag.begin.js
+      scope: punctuation.definition.tag.begin.js
       set:
         - jsx-meta
         - jsx-tag-attributes-top
